Add "0dev" Debian packaging
Start maintaining a development Debianization of MyPaint. If it works out,
this is intended to track the development master closely so that it's
always up to date.

Currently this can be used for personal ad-hoc builds, and is intended to
be used for building for the mypaint-testing PPA, and as a source of
patches for the Debian project's packaging efforts.